FIX: fixed hover jitter bug on card element by creating container around card
One of the visual glitch issues was the transition between
hover and the animation of it lifting the card up. When approaching
the element from the bottom, it would cause the animation to lift the
card up by -10px, but that would make the cursor technically not
touching the card which would stop the transition effect causing
visual jittering with the component. By surrounding the
card component with a div that upon :hover it will trigger the
li element by selecting it in css only on element hover, it will
let the transition play out. Now that there is a container, hovering
over the element from the bottom doesn't cause visual jitters.