FEAT: finished main game logic and score/highscore display
main game logic was held within the cardContainer.jsx
due to the fortnite API data and card shuffling interaction
being centralized there, but on the top level,
memoryGame.jsx controlled the score and triggered
re-renders of the main game if the score changed.
cardContainer.jsx re-renders the entire container when a
player clicks on a card, which shuffles the unordered list and adds
1 to the players score, then re-renders with the new order. I kept the data of each card remembered
by giving each fortniteData object a unique UUID key from cryto.randomUUID(),
which kept track of updated data within the fortniteData array, and makeCards()
would generate the cards in any order. Upon losing (clicking the same card more than once),
the game updates the highscore useState based on if the current player's score is higher and
resets the clickedOn object boolean values of each fortniteData object. The only thing I didn't
implement is resetting the position of the cards (unshuffle them). I plan to lastly
implement a <dialog> modal that will display if the player reaches 12 points.