FEAT: added useEffect() and useRef() logic to make a modal appear on score 12
using useRef() to create an out-of-render reference to the dialog
jsx element assigning it to the ref attribute, I made a useEffect that
checks on every score change if the score has reached 12 (the maximum score possible).
If it has reached 12, the player is shown a popup modal reward. Need
to style the modal, but a bug has been found in the shuffling algorithm that I
need to fix first