FIX: fixed bug in cardShuffle() in cardContainer.jsx that made the random index always 0
The random index that is being picked for the swap always was 0 because the
Math.floor() parentheses was only surrounding the Math.random() function which always made it 0. Moving the
parentheses of Math.floor() to cover the entire equation let the random index generate correctly. Cards
will now be shuffled in random order.